+++ x/group/keeper/msg_server.go
@@ -802,6 +802,7 @@ func (k Keeper) doTallyAndUpdate(ctx sdk.Context, p *group.Proposal, groupInfo g
 		} else {
 			p.Status = group.PROPOSAL_STATUS_REJECTED
 		}
+
 	}
 
 	return nil
@@ -837,7 +838,7 @@ func (k Keeper) Exec(goCtx context.Context, msg *group.MsgExec) (*group.MsgExecR
 			return nil, errorsmod.Wrap(err, "load group")
 		}
 
-		if err := k.doTallyAndUpdate(ctx, &proposal, groupInfo, policyInfo); err != nil {
+		if err = k.doTallyAndUpdate(ctx, &proposal, groupInfo, policyInfo); err != nil {
 			return nil, err
 		}
 	}
@@ -873,6 +874,16 @@ func (k Keeper) Exec(goCtx context.Context, msg *group.MsgExec) (*group.MsgExecR
 	// If proposal has successfully run, delete it from state.
 	if proposal.ExecutorResult == group.PROPOSAL_EXECUTOR_RESULT_SUCCESS {
 		if err := k.pruneProposal(ctx, proposal.Id); err != nil {
+			return nil, err
+		}
+
+		// Emit event for proposal finalized with its result
+		if err := ctx.EventManager().EmitTypedEvent(
+			&group.EventProposalPruned{
+				ProposalId:  proposal.Id,
+				Status:      proposal.Status,
+				TallyResult: &proposal.FinalTallyResult,
+			}); err != nil {
 			return nil, err
 		}
 	} else {